feat: 리뷰 반영 + 코드 컨벤션 반영
