+++ backend/selection/src/main/java/com/selection/SelectionApplication.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 @EnableJpaAuditing
 public class SelectionApplication {
 
-  public static void main(String[] args) {
-    SpringApplication.run(SelectionApplication.class, args);
-  }
+    public static void main(String[] args) {
+        SpringApplication.run(SelectionApplication.class, args);
+    }
 }
