+++ backend/selection/src/main/java/com/selection/domain/notice/Notice.java
@@ -19,24 +19,19 @@ import lombok.NoArgsConstructor;
 @Table(name = "NOTICES")
 public class Notice extends BaseEntity {
 
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  @Column(name = "NOTICE_ID")
-  private Long id;
+    @Column(nullable = false, length = 50)
+    private String title;
 
-  @Column(nullable = false, length = 50)
-  private String title;
+    @Column(nullable = false)
+    private String author;
 
-  @Column(nullable = false)
-  private String author;
+    @Lob
+    private String content;
 
-  @Lob
-  private String content;
-
-  @Builder
-  public Notice(String title, String author, String content) {
-    this.title = title;
-    this.author = author;
-    this.content = content;
-  }
+    @Builder
+    public Notice(String title, String author, String content) {
+        this.title = title;
+        this.author = author;
+        this.content = content;
+    }
 }
