+++ backend/selection/src/main/java/com/selection/domain/question/Question.java
@@ -21,21 +21,16 @@ import lombok.NoArgsConstructor;
 @Table(name = "QUESTIONS")
 public class Question extends BaseEntity {
 
-  @Id
-  @Column(name = "QUESTION_ID")
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
+    @Column(nullable = false, length = 100)
+    private String description;
 
-  @Column(nullable = false, length = 100)
-  private String description;
+    @ManyToOne
+    @JoinColumn
+    private Article article;
 
-  @ManyToOne
-  @JoinColumn(name = "ARTICLE_ID")
-  private Article article;
-
-  @Builder
-  public Question(String description, Article article) {
-    this.description = description;
-    this.article = article;
-  }
+    @Builder
+    public Question(String description, Article article) {
+        this.description = description;
+        this.article = article;
+    }
 }
