+++ backend/selection/src/main/java/com/selection/domain/tag/Tag.java
@@ -4,9 +4,6 @@ import com.selection.domain.BaseEntity;
 import com.selection.domain.article.Article;
 import javax.persistence.Column;
 import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
@@ -21,22 +18,17 @@ import lombok.NoArgsConstructor;
 @Table(name = "TAGS")
 public class Tag extends BaseEntity {
 
-  @Id
-  @Column(name = "TAG_ID")
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
+    @Column(nullable = false, length = 10)
+    private String name;
 
-  @Column(nullable = false, length = 10)
-  private String name;
+    @ManyToOne
+    @JoinColumn(name = "ARTICLE_ID")
+    private Article article;
 
-  @ManyToOne
-  @JoinColumn(name = "ARTICLE_ID")
-  private Article article;
-
-  @Builder
-  public Tag(String name, Article article) {
-    this.name = name;
-    this.article = article;
-  }
+    @Builder
+    public Tag(String name, Article article) {
+        this.name = name;
+        this.article = article;
+    }
 
 }
