+++ backend/selection/src/main/java/com/selection/dto/notice/NoticeResponse.java
@@ -7,18 +7,18 @@ import lombok.Getter;
 @Getter
 public class NoticeResponse {
 
-  private Long id;
-  private String title;
-  private String author;
-  private String content;
-  private LocalDateTime modifiedAt;
+    private Long id;
+    private String title;
+    private String author;
+    private String content;
+    private LocalDateTime modifiedAt;
 
 
-  public NoticeResponse(Notice notice) {
-    this.id = notice.getId();
-    this.title = notice.getTitle();
-    this.author = notice.getAuthor();
-    this.content = notice.getContent();
-    this.modifiedAt = notice.getModifiedAt();
-  }
+    public NoticeResponse(Notice notice) {
+        this.id = notice.getId();
+        this.title = notice.getTitle();
+        this.author = notice.getAuthor();
+        this.content = notice.getContent();
+        this.modifiedAt = notice.getModifiedAt();
+    }
 }
