+++ backend/selection/src/test/java/com/selection/controller/article/NoticeApiControllerTest.java
@@ -24,52 +24,53 @@ import org.springframework.web.filter.CharacterEncodingFilter;
 @SpringBootTest
 class NoticeApiControllerTest {
 
-  private MockMvc mockMvc;
+    private MockMvc mockMvc;
 
-  @Autowired
-  private WebApplicationContext ctx;
+    @Autowired
+    private WebApplicationContext ctx;
 
-  @Autowired
-  private NoticeReposiotry noticeReposiotry;
+    @Autowired
+    private NoticeReposiotry noticeReposiotry;
 
-  @BeforeEach
-  @DisplayName("테스트 목업 준비")
-  public void setUp() {
-    this.mockMvc = MockMvcBuilders.webAppContextSetup(ctx)
-        .addFilters(new CharacterEncodingFilter("UTF-8", true))  // UTF-8 인코딩 필터 추가
-        .alwaysDo(print())
-        .build();
+    @BeforeEach
+    @DisplayName("테스트 목업 준비")
+    public void setUp() {
+        this.mockMvc = MockMvcBuilders.webAppContextSetup(ctx)
+            .addFilters(new CharacterEncodingFilter("UTF-8", true))  // UTF-8 인코딩 필터 추가
+            .alwaysDo(print())
+            .build();
 
-    Notice notice = Notice.builder()
-        .title("공지사항 테스트")
-        .content("공지사항 테스트")
-        .author("관리자 1")
-        .build();
+        Notice notice = Notice.builder()
+            .title("공지사항 테스트")
+            .content("공지사항 테스트")
+            .author("관리자 1")
+            .build();
 
-    noticeReposiotry.save(notice);
-  }
+        noticeReposiotry.save(notice);
+    }
 
-  @AfterEach
-  @DisplayName("리소스 정리")
-  public void cleanUp() {
-    noticeReposiotry.deleteAll();
-  }
+    @AfterEach
+    @DisplayName("리소스 정리")
+    public void cleanUp() {
+        noticeReposiotry.deleteAll();
+    }
 
-  @Test
-  @DisplayName("공지사항 페이징 테스트")
-  public void getNotices() throws Exception {
-    final String PAGE = "1";
-    final String SIZE = "10";
-    final String DIRECTION = "ASC";
+    @Test
+    @DisplayName("공지사항 페이징 테스트")
+    public void getNotices() throws Exception {
+        final String PAGE = "1";
+        final String SIZE = "10";
+        final String DIRECTION = "ASC";
 
-    mockMvc.perform(
-        get("/api/notices")
-            .param("page", PAGE)
-            .param("size", SIZE)
-            .param("direction", DIRECTION)
-    )
-        .andExpect(status().isOk())
-        .andDo(print())
-        .andExpect(content().string(containsString("{\"id\":1,\"title\":\"공지사항 테스트\",\"author\":\"관리자 1\",\"content\":\"공지사항 테스트\",")));
-  }
+        mockMvc.perform(
+            get("/api/notices")
+                .param("page", PAGE)
+                .param("size", SIZE)
+                .param("direction", DIRECTION)
+        )
+            .andExpect(status().isOk())
+            .andDo(print())
+            .andExpect(content().string(containsString(
+                "{\"id\":1,\"title\":\"공지사항 테스트\",\"author\":\"관리자 1\",\"content\":\"공지사항 테스트\",")));
+    }
 }
